Fixed - Singleplayer end game issues

+++ src/main/java/ch/uzh/ifi/hase/soprafs21/service/GameService.java
@@ -267,9 +267,14 @@ public class GameService {
                 User user = userService.getUserByUserId(userId);
                 user.setInLobby(false);
             }
+        }            
+
+        // case 3: delete lobby only if multiplayer
+        UserMode uMode = game.getUserMode();
+        if (uMode.getName().equals("Multiplayer")){
+            lobbyService.deleteLobby(game.getLobbyId());
         }
 
-        lobbyService.deleteLobby(game.getLobbyId());
         gameRepository.delete(game);
         gameRepository.flush();
         userRepository.flush();
